added e2e test for endpoint to endpointslice migration
